backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sentence-transformers
faiss-cpu
requests
tqdm
scholarly
